api/requirements.txt
# FastAPI core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic v2
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
supabase==2.3.4
postgrest==0.13.2

# HTTP client for external APIs
httpx==0.26.0

# Authentication
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Stripe
stripe==8.1.0

# PDF generation
reportlab==4.0.9
weasyprint==60.2

# Email
resend==0.7.0

# Utilities
python-dateutil==2.8.2

# CORS
python-multipart==0.0.6
